+++ zanshinsdk/test_client.py
@@ -434,34 +434,6 @@ class TestClient(unittest.TestCase):
         )
 
     ###################################################
-    # Account API key
-    ###################################################
-
-    def test_iter_api_keys(self):
-        try:
-            next(self.sdk.iter_api_keys())
-        except StopIteration:
-            pass
-
-        self.sdk._request.assert_called_once_with("GET", "/me/apikeys")
-
-    def test_create_api_key(self):
-        name = "MyKey"
-
-        self.sdk.create_api_key(name)
-
-        self.sdk._request.assert_called_once_with(
-            "POST", "/me/apikeys", body={"name": name}
-        )
-
-    def test_delete_api_key(self):
-        api_key = "822f4225-43e9-4922-b6b8-8b0620bdb1e3"
-
-        self.sdk.delete_api_key(api_key)
-
-        self.sdk._request.assert_called_once_with("DELETE", f"/me/apikeys/{api_key}")
-
-    ###################################################
     # Organization
     ###################################################
